craete test cetnter controller
