+++ controller/RestEligibility.php
@@ -22,13 +22,14 @@ namespace oat\taoTestCenter\controller;
 use oat\taoTestCenter\model\eligibility\Eligibility;
 use oat\taoTestCenter\model\EligibilityService;
 use oat\taoDeliveryRdf\model\DeliveryAssemblyService;
-use oat\taoTestCenter\model\TestCenterService;
 use oat\tao\model\TaoOntology;
 
 /**
- * @OA\Info(title="TAO Test Center API", version="0.1")
+ * Class RestEligibility
+ * @package oat\taoTestCenter\controller
+ * @author Aleh Hutnikau, <hutnikau@1pt.com>
  */
-class RestEligibility extends \tao_actions_RestController
+class RestEligibility extends AbstractRestController
 {
 
     const PARAMETER_DELIVERY_ID = 'delivery';
@@ -344,18 +345,6 @@ class RestEligibility extends \tao_actions_RestController
     }
 
     /**
-     * Get test center resource from request parameters
-     * @return \core_kernel_classes_Resource
-     * @throws \common_exception_MissingParameter
-     * @throws \common_exception_NotFound
-     */
-    private function getTCFromRequest()
-    {
-        $testCenterUri = $this->getParameterFromRequest(self::PARAMETER_TEST_CENTER_ID);
-        return $this->getAndCheckResource($testCenterUri, TestCenterService::CLASS_URI);
-    }
-
-    /**
      * @return array
      * @throws \common_exception_RestApi
      * @throws \common_exception_NotFound
@@ -376,35 +365,5 @@ class RestEligibility extends \tao_actions_RestController
             throw new \common_exception_RestApi(__('`%s` parameter must be an array', self::PARAMETER_TEST_TAKER_IDS));
         }
         return $result;
-    }
-
-    /**
-     * @param $parameterName
-     * @return array|bool|mixed|null|string
-     * @throws \common_exception_MissingParameter
-     */
-    private function getParameterFromRequest($parameterName)
-    {
-        parse_str(file_get_contents("php://input"), $params);
-        $params = array_merge($params, $this->getRequestParameters());
-        if (!isset($params[$parameterName])) {
-            throw new \common_exception_MissingParameter(__('Missed `%s` parameter', $parameterName));
-        }
-        return $params[$parameterName];
-    }
-
-    /**
-     * @param $uri
-     * @param null $class
-     * @return \core_kernel_classes_Resource
-     * @throws \common_exception_NotFound
-     */
-    private function getAndCheckResource($uri, $class = null)
-    {
-        $resource = $this->getResource($uri);
-        if (!$resource->exists() || ($class !== null && !$resource->hasType($this->getClass($class)))) {
-            throw new \common_exception_NotFound(__('Resource with `%s` uri not found', $uri));
-        }
-        return $resource;
     }
 }
